Fix parser's handling of implicit tokens
This ensures EOF is considered as a terminator for implicit commas. It
also fixes a bug where map/array members terminated by the '}' or ']'
(and not a comma or newline/EOF) were not added to the list of members.

This commit _does_ add some weird handling of EOF in the `eat` function
which I should probably fix somewhere else, or maybe add a comment to
make sure it's known about in the future.

Basically, eating an EOF doesn't take it off the token iterator like
eating other tokens does. Expecting an EOF _does_ still take it off the
iterator, though. This sounds like it could cause confusion.