Add Hjson parser
This parser is based on this excellent article by matklad:
<https://matklad.github.io/2023/05/21/resilient-ll-parsing-tutorial.html>.

The parser has no tests at the moment, mostly because they're quite a
lot of effort to write.

It's difficult to know what bases to cover, and the naive way of writing
tests (specifying the expected tree) gets super super verbose.